update executor proto; fix operation env var for executor and merkletree
